+++ edison-cache/src/main/java/de/otto/edison/cache/configuration/CacheRegistry.java
@@ -3,6 +3,9 @@ package de.otto.edison.cache.configuration;
 
 import com.github.benmanes.caffeine.cache.Cache;
 
+/**
+ * @Beta
+ */
 public interface CacheRegistry {
 
     void registerCache(CaffeineCacheConfig config, Cache cache);