+++ edison-cache/src/main/java/de/otto/edison/cache/controller/CacheInfoController.java
@@ -1,12 +1,13 @@
 package de.otto.edison.cache.controller;
 
-import de.otto.edison.annotations.Beta;
+import com.github.benmanes.caffeine.cache.Cache;
 import de.otto.edison.cache.configuration.CaffeineCacheConfig;
 import de.otto.edison.navigation.NavBar;
-import de.otto.edison.navigation.NavBarItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.endpoint.CachePublicMetrics;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.cache.CacheManager;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,9 +20,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.google.common.base.CaseFormat.LOWER_CAMEL;
-import static com.google.common.base.CaseFormat.LOWER_UNDERSCORE;
-import static de.otto.edison.navigation.NavBarItem.*;
+import static de.otto.edison.navigation.NavBarItem.bottom;
+import static de.otto.edison.navigation.NavBarItem.navBarItem;
 import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.toList;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
@@ -33,8 +33,7 @@ import static org.springframework.web.bind.annotation.RequestMethod.GET;
  * @since 0.76.0
  */
 @Controller
-@Beta
-@ConditionalOnProperty(name = "edison.cache.web.controller.enabled", havingValue = "true", matchIfMissing = true)
+@ConditionalOnBean(CaffeineCacheConfig.class)
 public class CacheInfoController {
 
     @Autowired
@@ -88,12 +87,32 @@ public class CacheInfoController {
             final String name = m.getName().substring("cache.".length());
             int pos = name.indexOf('.');
             final String cacheName = name.substring(0, pos);
-            final String metricName = LOWER_UNDERSCORE.to(LOWER_CAMEL, name.substring(pos+1).replace(".", "_").replace("-", "_")) ;
+            final String metricName = toCamelHumps(normalizeDotsAndDashes(name, pos)) ;
             if (!cacheInfos.containsKey(cacheName)) {
                 cacheInfos.put(cacheName, new CacheInfo(cacheName));
             }
             cacheInfos.get(cacheName).setMetric(metricName, m.getValue());
         });
         return cacheInfos;
+    }
+
+    private String normalizeDotsAndDashes(final String name, final int pos) {
+        return name.substring(pos+1).replace(".", "_").replace("-", "_");
+    }
+
+    private String toCamelHumps(final String input) {
+        final StringBuilder sb = new StringBuilder();
+        boolean isFirst = true;
+        for( String word : input.split("_") )
+        {
+            if (isFirst) {
+                sb.append(word.substring(0, 1));
+                isFirst = false;
+            } else {
+                sb.append(word.substring(0, 1).toUpperCase());
+            }
+            sb.append( word.substring(1).toLowerCase() );
+        }
+        return sb.toString();
     }
 }